refactor: Rename "InstitutionsDetails" to "ViewInstitutionProfile" to better align with all the profile forms we are developing
