+++ src/pages/Filing/UpdateFinancialProfile/index.tsx
@@ -22,7 +22,7 @@ import {
 } from 'pages/Filing/UpdateFinancialProfile/types';
 import InputEntry from 'pages/ProfileForm/Step1Form/InputEntry';
 
-import { Controller, useForm } from 'react-hook-form';
+import { Controller as FormController, useForm } from 'react-hook-form';
 
 // TODO: Decide on properties to inherit
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
@@ -105,7 +105,7 @@ function UpdateFinancialProfile(properties: Properties): JSX.Element {
                 };
                 return (
                   <ListItem key={option.id}>
-                    <Controller
+                    <FormController
                       render={({ field }) => (
                         // TS error should be fixed in DSR Repo
                         // @ts-expect-error
