+++ src/pages/Filing/ViewUserProfile/AssociatedInstitutions.tsx
@@ -1,8 +1,8 @@
 import { Link } from 'components/Link';
 import { Heading, List, Paragraph, WellContainer } from 'design-system-react';
-import type { InstitutionDetailsApiType } from 'pages/Filing/InstitutionDetails/institutionDetails.type';
-import { AssociatedInstitution } from '../../../components/AssociatedInstitution';
+import type { InstitutionDetailsApiType } from 'pages/Filing/ViewInstitutionProfile/institutionDetails.type';
 import { sblHelpLink } from 'utils/common';
+import { AssociatedInstitution } from '../../../components/AssociatedInstitution';
 
 export default function AssociatedInstitutions({
   data: associatedInstitutions,
