+++ src/utils/constants.ts
@@ -1,3 +1,4 @@
 // Needed due to ESLint rule: @typescript-eslint/no-magic-numbers
 
 export const One = 1;
+export const Five = 5;
